# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pearlis Luxury Jewelry API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: material
          schema: { type: string }
        - in: query
          name: minPrice
          schema: { type: number }
        - in: query
          name: maxPrice
          schema: { type: number }
        - in: query
          name: sort
          schema:
            type: string
            enum: [latest, price_asc, price_desc, trending, rating]
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 12 }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # CATEGORIES
...
  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Add a review
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # CART
...
  /wishlist/{productId}:
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove product from wishlist
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # BLOGS
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    Product:
      type: object
      required: [id, name, price, category, images, stock, isNew, isTrending, rating, reviewCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
          nullable: true
        category:
          type: string
        categoryId:
          type: integer
        material:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isNew:
          type: boolean
        isTrending:
          type: boolean
        isFeatured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ProductListResponse:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CreateProductBody:
      type: object
      required: [name, price, category, images, stock]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
          nullable: true
        category:
          type: string
        categoryId:
          type: integer
        material:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isNew:
          type: boolean
        isTrending:
          type: boolean
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string
...
    Review:
      type: object
      required: [id, productId, rating, comment, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
          nullable: true
        userName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
    CreateReviewBody:
      type: object
      required: [rating, comment]
      properties:
        rating:
          type: integer
        comment:
          type: string
        userName:
          type: string
...